requirements.txt
Flask==0.12
Flask-Script==0.4.0
python-telegram-bot==9.0.0
transitions==0.6.3
spotipy==2.4.4